Use new resumption protocol.
Suspensions may not be resumed by a process that has any object locked.
For procedures that need to know whether a suspension can be resumed
before it can unlock an object, the procedure WITH-SUSPENSION-CLAIMED
gives them this information by claiming suspensions so that any other
processes trying to use them will spin or block until the suspension
has either been resumed or disclaimed.  MAYBE-RESUME may no longer be
called when any object is locked, and is simply a special case of
WITH-SUSPENSION-CLAIMED.  Note that only one suspension may be claimed
at any one time.  This is a limitation that is hard to lift -- and I
suspect that it is a limitation intrinsic to CML.  I don't know what
might require the limitation to be lifted, though.

In the process, simplify the implementations of channels, mailboxes,
placeholders, and (internal) condition variables.