Use new primitive suspension protocol in primitive.scm.
Move suspension abstraction into primitive.scm, using a lower-level
`suspender' abstraction that is easier to implement, enabling the
complexity of suspensions to be implemented only once and isolated.
A suspender is an object with a lock and with which a process can
suspend until another process resumes it with a value.